made one final tweak to sub task delete for alert

+++ app/Controllers/SubTasksController.js
@@ -21,7 +21,12 @@ export default class SubTasksController {
     }
 
     deleteSubTasks(subTaskId) {
-        window.confirm('Are you sure you want to delete this?')
-        subTasksService.deleteSubTask(subTaskId)
+        let result = window.confirm('Are you sure you want to delete this?')
+        if (result) {
+            subTasksService.deleteSubTask(subTaskId)
+        }
+        else {
+            return;
+        }
     }
 }
